npm workspace 설정 & 약간의 리팩토링 (#19)
npm workspace 설정과 client의 App.js에서 컴포넌트로 쪼개서 약간 리팩토링했습니다.
