+++ client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "react-client",
+  "name": "client",
   "version": "0.1.0",
   "private": true,
   "proxy": "http://localhost:8080",
@@ -19,9 +19,9 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "lint": "./node_modules/.bin/eslint .",
-    "lint:fix": "./node_modules/.bin/eslint . --fix",
-    "pre-commit" : "./node_modules/.bin/lint-staged"
+    "lint": "../node_modules/.bin/eslint .",
+    "lint:fix": "../node_modules/.bin/eslint . --fix",
+    "pre-commit" : "../node_modules/.bin/lint-staged"
   },
   "eslintConfig": {
     "extends": [
