+++ client/src/components/ViewContent.js
