+++ client/src/components/WriteContent.js
