+++ client/src/route/ContentsView.js
@@ -1,0 +1,14 @@
+import React from "react";
+import ContentsList from "../components/ContentsList";
+import ContentTitle from "../components/ContentTitle";
+
+const ContentsView = () => {
+  return (
+    <div>
+      <ContentTitle />
+      <ContentsList />
+    </div>
+  );
+};
+
+export default ContentsView;
