+++ package.json
@@ -6,5 +6,13 @@
   },
   "scripts": {
     "prepare": "husky install"
+  },
+  "private": true,
+  "workspaces": [
+    "client",
+    "server"
+  ],
+  "dependencies": {
+    "lint-staged": "^11.2.0"
   }
 }
