--- react_server/goorm.manifest
@@ -1,1 +1,0 @@
-{"author":"6f3709af5343f1d31fa13a9b8df2a99bdbd96071ec8066b00c33c97cc3609540_yr9cz_naver","name":"react_server","type":"react","detailedtype":"react","description":"REACT 공부를 위한 컨테이너","date":"2021-07-04T13:08:22.000Z","plugins":{"goorm.plugin.react":[{"name":"react"}]},"is_user_plugin":false,"storage":"container","project_domain":[{"id":"6f3709af5343f1d31fa13a9b8df2a99bdbd96071ec8066b00c33c97cc3609540_yr9cz_naver","url":"react-ohbae-bzomu.run.goorm.io","port":"3000"}],"show_preview_btn":true,"author_email":"0176778602@naver.com","author_name":"hyunw3","ignore_patterns":[],"visibility":1}
